backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
djangorestframework-simplejwt
dj-rest-auth[with-social]
django-allauth[socialaccount]
requests